+++ modules/lambdas_secondary/main.tf
@@ -1,0 +1,55 @@
+## =============================================================================
+## Reconciliation Lambdas Definitions and Resources
+## =============================================================================
+
+# post_to_queue_and_trigger_step_function - Receives an events from an SQS queue, translates to get_current_archive_list's input format, sends it to another queue, then triggers the internal report step function.
+# ==============================================================================
+resource "aws_lambda_function" "post_to_queue_and_trigger_step_function" {
+  ## REQUIRED
+  function_name = "${var.prefix}_post_to_queue_and_trigger_step_function"
+  role          = var.restore_object_role_arn  # todo: Specialized roles. ORCA-316
+
+  ## OPTIONAL
+  description      = "Receives an events from an SQS queue, translates to get_current_archive_list's input format, sends it to another queue, then triggers the internal report step function."
+  filename         = "${path.module}/../../tasks/post_to_queue_and_trigger_step_function/post_to_queue_and_trigger_step_function.zip"
+  handler          = "post_to_queue_and_trigger_step_function.handler"
+  memory_size      = var.orca_reconciliation_lambda_memory_size
+  runtime          = "python3.9"
+  source_code_hash = filebase64sha256("${path.module}/../../tasks/post_to_queue_and_trigger_step_function/post_to_queue_and_trigger_step_function.zip")
+  tags             = var.tags
+  timeout          = var.orca_reconciliation_lambda_timeout
+
+  vpc_config {
+    subnet_ids         = var.lambda_subnet_ids
+    security_group_ids = [var.vpc_postgres_ingress_all_egress_id]
+  }
+
+  environment {
+    variables = {
+      PREFIX = var.prefix
+      STEP_FUNCTION_ARN = var.orca_sfn_internal_reconciliation_workflow_arn,
+      TARGET_QUEUE_URL  = var.orca_sqs_internal_report_queue_id,
+    }
+  }
+}
+
+
+# todo: Populate queue with s3 events ORCA-372 https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/s3_bucket_notification
+# Additional resources needed by post_to_queue_and_trigger_step_function
+# ------------------------------------------------------------------------------
+resource "aws_lambda_event_source_mapping" "post_to_queue_and_trigger_step_function_event_source_mapping" {
+  event_source_arn = var.orca_sqs_s3_inventory_queue_arn
+  function_name    = aws_lambda_function.post_to_queue_and_trigger_step_function.arn
+}
+
+# Permissions to allow SQS trigger to invoke lambda
+resource "aws_lambda_permission" "post_to_queue_and_trigger_step_function_allow_sqs_trigger" {
+  ## REQUIRED
+  action        = "lambda:InvokeFunction"
+  function_name = aws_lambda_function.post_to_queue_and_trigger_step_function.function_name
+  principal     = "sqs.amazonaws.com"
+
+  ## OPTIONAL
+  statement_id = "AllowExecutionFromSQS"
+  source_arn   = var.orca_sqs_s3_inventory_queue_arn
+}